requirements.txt
appdirs==1.4.3
attrs==19.3.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
idna==2.9
isodate==0.6.0
lxml==4.5.0
pytz==2019.3
requests==2.23.0
requests-toolbelt==0.9.1
six==1.14.0
suds-community==0.8.4
urllib3==1.25.8
xmltodict==0.12.0
zeep==3.4.0
